iteration

+++ data-features/feature-api/src/commonMain/kotlin/com/grippo/data.features.api/exercise/example/models/ExerciseExampleComponents.kt
@@ -8,18 +8,18 @@ public sealed class ExerciseExampleComponents {
 
     public data class BodyOnly(
         val required: Boolean,
-        val multiplier: Double
+        val multiplier: Float
     ) : ExerciseExampleComponents()
 
     public data class BodyAndExtra(
         val bodyRequired: Boolean,
-        val bodyMultiplier: Double,
+        val bodyMultiplier: Float,
         val extraRequired: Boolean
     ) : ExerciseExampleComponents()
 
     public data class BodyAndAssist(
         val bodyRequired: Boolean,
-        val bodyMultiplier: Double,
+        val bodyMultiplier: Float,
         val assistRequired: Boolean
     ) : ExerciseExampleComponents()
 }
